Update method of point detection
Significantly changes the way that points are detected. Wherever there has been detected an island of pixels that is greater than or equal to half the size of a point, put a point there. Then, subtract from the original total the size of the point. Continue until the total size is less than half. Essentially, we need to populate each island until it is at capacity. My idea for the algorithm is the maximize the distance from 'water'.  So first place a point in the middle of the island. Then where that point is placed, water will be filled in.